evenly.spaced {spatstat.utils}	R Documentation

Determine Whether a Vector is Evenly Spaced and Increasing

Description

Determines whether the entries in a numeric vector are evenly spaced and increasing.

Usage

evenly.spaced(x, tol = 1e-07)

Arguments

x
Numeric vector.

tol
Relative tolerance.

Details

The result is TRUE if x is an increasing sequence in which the successive differences diff(x) are all equal to one another (within the specified relative tolerance), and FALSE otherwise.

Value

A single logical value.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

Examples

  evenly.spaced(seq(0, 1, length=4))

[Package spatstat.utils version 3.1-2 Index]