geomseq {spatstat.utils}	R Documentation

Geometric Sequence

Description

Generate a geometric sequence between two endpoints. The sequence is equally spaced on a logarithmic scale.

Usage

geomseq(from, to, length.out)

Arguments

from
Starting value. A positive number.

to
Ending value. A positive number.

length.out
Number of elements in the sequence. A positive integer.

Details

This is a wrapper for seq.default which generates a geometric sequence between the two endpoints.

Value

Numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

seq.default

Examples

  geomseq(1, 32, length.out=6)

[Package spatstat.utils version 3.1-2 Index]