splat {spatstat.utils}	R Documentation

Print Text Within Margins

Description

Prints a given character string or strings inside the text margin specified by options("width"). Indents the text if required.

Usage

splat(..., indent = 0)

Arguments

...
Character strings, or other arguments acceptable to paste.

indent
Optional. Indentation of the text. Either an integer specifying the number of character positions by which the text should be indented, or a character string whose length determines the indentation.

Details

splat stands for ‘split cat’.

The command splat(...) is like cat(paste(...)) except that the output will be split into lines that can be printed within the current text margin specified by getOption("width").

The arguments ... are first combined into a character vector using paste. Then they are split into words separated by white space. A newline will be inserted whenever the next word does not fit in the available text area. (Words will not be broken, so the text margin could be exceeded if any word is longer than getOption("width")).

If any argument is a vector, each element of the vector is treated as a separate line. Existing newline characters in ... are also respected.

Value

Null.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  op <- options(width=20)
  splat("There is more than one way to skin a cat.")
  splat("There is more than one", "way to skin a cat.", indent=5)

  options(width=10)
  splat("The value of pi is", pi)
  splat("The value of pi is", signif(pi))
  options(op)

[Package spatstat.utils version 3.1-2 Index]